refactor(gateway): share unavailable method responses
Keep the same catch boundaries, error codes, messages, redaction, and
node-invoke provenance while routing default failures through one mapper.
DM approval keeps its original control flow and pairing-adapter rereads.

Proof: 1921 focused tests pass; core typecheck and independent review pass.
