+++ src/gateway/server-methods/nodes.helpers.ts
@@ -1,22 +1,11 @@
-// Node method helpers centralize unavailable responses, safe JSON parsing,
-// and node-invoke error mapping.
+// Node method helpers centralize JSON parsing and node-invoke error mapping.
 import { normalizeOptionalString } from "@openclaw/normalization-core/string-coerce";
 import {
   ErrorCodes,
   errorShape,
 } from "../../../packages/gateway-protocol/src/schema/error-codes.js";
-import { formatForLog } from "../ws-log.js";
 import type { RespondFn } from "./types.js";
 export { parseGatewayPayload } from "../server-json.js";
-
-/** Converts thrown node-handler failures into `UNAVAILABLE` protocol errors. */
-export async function respondUnavailableOnThrow(respond: RespondFn, fn: () => Promise<void>) {
-  try {
-    await fn();
-  } catch (err) {
-    respond(false, undefined, errorShape(ErrorCodes.UNAVAILABLE, formatForLog(err)));
-  }
-}
 
 /** Narrows successful node invoke results or responds with the node error details. */
 export function respondUnavailableOnNodeInvokeError<T extends { ok: boolean; error?: unknown }>(
