+++ src/gateway/server-methods/nodes.pending.ts
@@ -20,8 +20,8 @@ import {
   type PendingNodeAction,
 } from "../node-runtime-state.js";
 import { nodeInvokePolicy } from "./nodes-policy.js";
-import { respondUnavailableOnThrow } from "./nodes.helpers.js";
 import { respondPairingChanged } from "./nodes.shared.js";
+import { respondUnavailableOnThrow } from "./response.js";
 import type { GatewayRequestHandlers } from "./types.js";
 import { assertValidParams } from "./validation.js";
