+++ src/gateway/server-methods/voicewake.ts
@@ -2,18 +2,16 @@
 import { ErrorCodes, errorShape } from "../../../packages/gateway-protocol/src/index.js";
 import { loadVoiceWakeConfig, setVoiceWakeTriggers } from "../../infra/voicewake.js";
 import { normalizeVoiceWakeTriggers } from "../server-utils.js";
-import { formatForLog } from "../ws-log.js";
+import { respondUnavailableOnThrow } from "./response.js";
 import type { GatewayRequestHandlers } from "./types.js";
 
 /** Gateway request handlers for reading and updating voice wake triggers. */
 export const voicewakeHandlers: GatewayRequestHandlers = {
   "voicewake.get": async ({ respond }) => {
-    try {
+    await respondUnavailableOnThrow(respond, async () => {
       const cfg = await loadVoiceWakeConfig();
       respond(true, { triggers: cfg.triggers });
-    } catch (err) {
-      respond(false, undefined, errorShape(ErrorCodes.UNAVAILABLE, formatForLog(err)));
-    }
+    });
   },
   "voicewake.set": async ({ params, respond, context }) => {
     if (!Array.isArray(params.triggers)) {
@@ -24,15 +22,13 @@ export const voicewakeHandlers: GatewayRequestHandlers = {
       );
       return;
     }
-    try {
+    await respondUnavailableOnThrow(respond, async () => {
       const triggers = normalizeVoiceWakeTriggers(params.triggers);
       // Persist the normalized trigger list before broadcasting so connected
       // nodes and future gateway starts observe the same wake phrases.
       const cfg = await setVoiceWakeTriggers(triggers);
       context.broadcastVoiceWakeChanged(cfg.triggers);
       respond(true, { triggers: cfg.triggers });
-    } catch (err) {
-      respond(false, undefined, errorShape(ErrorCodes.UNAVAILABLE, formatForLog(err)));
-    }
+    });
   },
 };
